requirements.txt
django>=5.0
django-htmx
intuit-oauth>=1.2.6
python-quickbooks>=0.9.12
python-dotenv>=1.0.0
